requirements.txt
Django==3.2.12
django-hosts==5.1
django-extensions==3.1.5
django-cors-headers==3.11.0
django-tables2==2.4.1
django-filter==21.1
Fabric3==1.14.post1
django-widget-tweaks==1.4.12
jsonfield==3.1.0
flake8==4.0.1
ipython==8.4.0
